TAG-298 Mer renaming
