+++ src/main/java/no/nav/tag/kontakt/oss/fylkesinndelingMedNavEnheter/Bydel.java
@@ -1,4 +1,4 @@
-package no.nav.tag.kontakt.oss.fylkesinndelingMedNavenheter.integrasjon;
+package no.nav.tag.kontakt.oss.fylkesinndelingMedNavEnheter;
 
 import lombok.AllArgsConstructor;
 import lombok.EqualsAndHashCode;
