+++ src/main/java/no/nav/tag/kontakt/oss/fylkesinndelingMedNavEnheter/NavEnhet.java
@@ -1,0 +1,10 @@
+package no.nav.tag.kontakt.oss.fylkesinndelingMedNavEnheter;
+
+import lombok.EqualsAndHashCode;
+import lombok.Value;
+
+@EqualsAndHashCode
+@Value
+public class NavEnhet {
+    private String enhetNr;
+}
